+ added Fleet and coordinates model for Db and Logic
