+++ GamePortal/AliaksNad.Battleship.Data/Contexts/UsersContexts.cs
@@ -30,24 +30,4 @@ namespace AliaksNad.Battleship.Data.Contexts
             modelBuilder.Configurations.Add(new UserDbConfiguration());
         }
     }
-
-    public sealed class FleetContexts : DbContext
-    {
-        public FleetContexts(ILogger logger)
-        {
-            //Database.SetInitializer<UsersContexts>(new MigrateDatabaseToLatestVersion<UsersContexts, Configuration>());
-            Database.SetInitializer<FleetContexts>(new DropCreateDatabaseAlways<FleetContexts>());
-            Database.Log = msg => logger.Warning(msg);
-        }
-
-        public DbSet<FleetDb> FleetCoordinates { get; set; }
-
-        public DbSet<CoordinatesDb> Coordinates { get; set; }
-
-        protected override void OnModelCreating(DbModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-            modelBuilder.Configurations.Add(new FleetDbConfiguration());
-        }
-    }
 }
