+++ GamePortal/AliaksNad.Battleship.Logic/BattleshipLogicDIModule.cs
@@ -17,7 +17,7 @@ namespace AliaksNad.Battleship.Logic
     {
         public override void Load()
         {
-            Mapper.Initialize(cfg => cfg.AddProfiles(typeof(UserProfile), typeof(FleetProfile), typeof(CoordinatesProfile)));
+            Mapper.Initialize(cfg => cfg.AddProfiles(typeof(UserProfile)));
             var mapper = Mapper.Configuration.CreateMapper();
 
             this.Bind<IMapper>().ToConstant(mapper);
