+++ GamePortal/AliaksNad.Battleship.Logic/Models/CoordinatesDto.cs
@@ -1,0 +1,11 @@
+﻿namespace AliaksNad.Battleship.Logic.Models
+{
+    public class CoordinatesDto
+    {
+        public int Id { get; set; }
+
+        public int X { get; set; }
+
+        public int Y { get; set; }
+    }
+}
