+++ GamePortal/AliaksNad.Battleship.Logic/Profiles/FleetProfile.cs
@@ -1,16 +1,14 @@
 ﻿using AliaksNad.Battleship.Data.Models;
 using AliaksNad.Battleship.Logic.Models;
 using AutoMapper; // or Mapster
-using System.Security.Cryptography.X509Certificates;
 
 namespace AliaksNad.Battleship.Logic.Profiles
 {
-    class UserProfile : Profile
+    class FleetProfile : Profile
     {
-        public UserProfile()
+        public FleetProfile()
         {
-            CreateMap<UserDb, UserDto>()
-                .ForMember(x => x.Password, opt => opt.Ignore())
+            CreateMap<FleetDb, FleetDto>()
                 .ReverseMap();
         }
     }
