+++ GamePortal/AliaksNad.Battleship.Logic/Services/IGameService.cs
@@ -17,6 +17,6 @@ namespace AliaksNad.Battleship.Logic.Services
         /// </summary>
         /// <param name="coordinates">Enemy coordinates.</param>
         /// <returns></returns>
-        bool CheckHit(Coordinates coordinates);
+        bool CheckHit(CoordinatesCheck coordinates);
     }
 }
